requirements.txt
alembic==1.7.7
blinker==1.4
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
coverage==6.3.2
Faker==13.3.4
Flask==2.1.1
Flask-CORS==3.0.10
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
lesscpy==0.15.0
mypy==0.940
ply==3.11
psycopg2-binary==2.9.3
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
smartypants==2.0.1
SQLAlchemy==1.4.35
text-unidecode==1.3
urllib3==1.26.9
URLObject==2.4.3
Werkzeug==2.1.1
